hacked rock processing to help two rock fighting
However, one rock is not as good.  And there's lots of debug and cleanup,
The rock code needs to be smarter so it culsters pixels and understands
multiple rocks.  It's a big hack without it.  But I'm getting burned
our hacking this code and we are so close to having it good eougbh to find all rocks
now...